+++ functions.php
@@ -194,6 +194,17 @@ function piptheme_get_parent_stylesheet_uri() {
 	}
 }
 
+/* from https://dotlayer.com/what-is-migrate-js-why-and-how-to-remove-jquery-migrate-from-wordpress/ */
+function remove_jquery_migrate($scripts) {
+  if (! is_admin() && isset($scripts->registered['jquery'])) {
+    $script = $scripts->registered['jquery'];
+    if ($script->deps) {
+      $script->deps = array_diff($script->deps, array('jquery-migrate'));
+    }
+  }
+}
+add_action('wp_default_scripts', 'remove_jquery_migrate');
+
 /* from https://kinsta.com/knowledgebase/disable-emojis-wordpress/#disable-emojis-code */
 /**
  * Disable the emoji's